reworking training pair/neighbor list.  updated test_pairlist.py
