+++ modelforge/potential/painn.py
@@ -9,7 +9,8 @@ import torch.nn as nn
 from loguru import logger as log
 from openff.units import unit
 
-from .models import NNPInputTuple, PairlistData
+from modelforge.dataset.dataset import NNPInputTuple
+from modelforge.potential.neighbors import PairlistData
 from .utils import DenseWithCustomDist
