+++ modelforge/potential/sake.py
@@ -8,7 +8,9 @@ import torch
 import torch.nn as nn
 from loguru import logger as log
 
-from .models import NNPInputTuple, PairlistData
+from modelforge.dataset.dataset import NNPInputTuple
+from modelforge.potential.neighbors import PairlistData
+
 from .utils import DenseWithCustomDist, scatter_softmax
 from .representation import PhysNetRadialBasisFunction
